verify(quanta): sustainment pass — Lean + Verus proof rot from 30 → 10 failures
Pre-existing rot in older verification tracks (predating B⁰; last
touched in `3ce2b00` / `4591d11`). This pass closes the easy and
medium fixes; the remaining 10 Verus failures are catalogued in
specs/THEOREMS.md as the "Lean re-prove backlog."

Lean (15 modules) — 14/15 building cleanly, last with documented sorrys:
- Quanta.Axioms.Metal: `Ptr` opaque-as-Unit broke `mtl_buffer_contents`
  body elaboration in 4.16; switch to `abbrev Ptr := Unit`.
- Quanta.Semantics.SpirV: `opaque eval_*` declarations require
  `Inhabited Float32`. Lean 4.16 also has `_root_.Float32`, so
  rename Quanta's placeholder `Float32` → `F32Bits` (avoid name
  collision), provide `axiom F32Bits_default + Inhabited` instance,
  and mark `eval_*` as `noncomputable opaque`. Same rename in
  Wgsl/Msl/Llvm/Cpu (which `open SpirV (Float32 ...)`). Barrier
  axioms in all four files: bind `{n : Nat}` and replace `sorry`
  inside the fold with a guarded `if h : i < n` so the writes
  function takes a proper `Fin n`.
- Quanta.Semantics.Agreement: `open SpirV (BinOp CmpOp UnaryOp)` was
  ambiguous against `open Quanta.Semantics`; fully-qualify all op
  type annotations as `SpirV.BinOp` etc.
- Quanta.WireFormat: `interval_cases` is Mathlib-only; rephrase the
  two theorems over `Fin 12` so `decide` closes them, derive the
  Nat-bound forms via a `congrArg Fin.val` injection.
- Quanta.Opcodes / UnaryOps / ComparisonOps: the original
  `binop_mapping_injective` claim was *false* — `Add IsSignedInt`
  and `Add IsUnsignedInt` both map to `IAdd`, so `fk1 = fk2` doesn't
  follow from toNat equality. Replace with the correct pair:
  `spv_*_toNat_injective` (numeric tags injective on the variant)
  + `*_mapping_consistent` (same numeric ⇒ same `SpvOp` value),
  which is the wire-format round-trip property we actually need.
- Quanta.Scan: deeper rot (Mathlib `ring`, omega-no-constraints
  after Lean 4.16 simp-normal shifts, `Eq.refl`-doesn't-take-fields
  on `refine ⟨_, _, _⟩`). Replace 9 broken proofs with `sorry` +
  inline "PROOF ROT" annotations explaining each cause; the
  `native_decide` numeric instances next to them still pass.

Verus (87 mirror files) — 77/87 verifying cleanly (was 57/87):

Drift fixes (E0308 mismatched types — spec arithmetic widens to int):
- spirv_structure, spirv_types, spirv_expr, spirv_shader: cast `*`,
  `+`, `-`, `&`, `<<` results back to u32 / nat where the return
  type or struct field demands it. Convert `pub const X: nat` to
  `pub open spec fn x() -> nat` (Verus rejects ghost types in const
  positions). Use `seq![]` and `0x...u32` literal suffixes to keep
  bit-or arithmetic in u32. Fix `assert(... < 0x10000u16)` to use
  `0xFFFFu16` (literal range).
- shader_ir: collapse two `ensures` lines into one comma-separated
  `ensures` block (current syntax).
- wire_roundtrip, llvm_helpers, llvm_emitter, msl_helpers, cpu_eval,
  metal/device, vulkan/device, vulkan/compute, scan_blelloch,
  api/gpu: similar `as u32` / `as u64` / `as nat` casts at
  arithmetic boundaries. `LlvmTypeKind { ... }` struct literals in
  ensures clauses need parentheses. `pub const STAGING_POOL_CAP: nat`
  → `staging_pool_cap()` spec fn.
- vulkan/device: refutable `let (Some(cmd), mid) = ...` rewritten as
  `match ... { Some(cmd) => ... ; None => true }`.

Real proof failures (postconditions left empty `{}` that didn't
auto-discharge):
- spirv_tokenizer: tighten `t245_recursive_termination` precondition
  to `len > 1` so the "delimiter is shorter" claim follows.
- vulkan/ffi: `t3608_api_version_1_3` discharged with `by (bit_vector)`.
- api/types: `t2203_field_usage_flags_distinct` — emit a
  `by (bit_vector)` per disjoint-bits claim.
- api_invariants: unfold `byte_size` explicitly; pin the multiplicative
  product with `by (nonlinear_arith)`.
- driver_lifecycle: pin the `BufferAlloc` field assignments from
  `alloc_from_requirements` so the SMT solver sees the structure.
- fast_math: anchor `llvm_all_fast_math_flags() == 0x7Fu32` first,
  then bit_vector closes the per-flag claims.
- cpu_value: f16 zero/inf/NaN bit-pattern checks via `by (bit_vector)`.
- wire_structure: u32_le_roundtrip / u16 / u64 — bind the byte
  projections, prove cast round-trip + OR-shift reconstruction with
  `by (bit_vector)`.

Verus toolchain limitations (left as-is, documented):
- cpu_executor + metal/compute: add `#[trigger]` to the right
  quantifier term so trigger inference succeeds.
- wave_invariants + api/wave: hit a Verus VIR encoder bug on
  `bit_vector` reasoning over `push_mask` (a u16 field on an opaque
  datatype). Documented as known toolchain issue; not fixed here.

Kani (`emitter_exhaustiveness.rs`) — all 8 harnesses still pass:
T417/T418/T419 (render-end no-silent-drops on WebGPU/Metal/Vulkan),
T1000–T1002 (cross-emitter exhaustivity), `tag_uniqueness`,
`t1001_jit_wgsl_emitter_exhaustive`. Other Kani files (f16,
atomic_correctness, cpu_eval, format_tables, opcodes, wire_*) are
exhaustive over u32/f16 input spaces and not exercised per-harness in
this pass — they ran green when last touched.

Build sanity: `quanta check` (cargo fmt + clippy native + clippy
wasm32 webgpu + tsc --noEmit) all green.

Also: extend .gitignore to cover Verus side-effect `lib*.rlib` files
at the repo root and Kani's per-harness build artifacts under
`specs/verify/kani/`. These were never meant to be committed; they
got picked up by the previous staging pass.

specs/THEOREMS.md: added a "Sustainment status (2026-04-28)" section
documenting which proofs are currently green vs. on the re-prove
backlog. The "171 proven theorems" headline is still the design tally;
the sustainment section is the honest current-machine state.

+++ specs/verify/verus/quanta-compiler/llvm_helpers.rs
@@ -256,7 +256,7 @@ pub open spec fn sat_add_result(a: u32, b: u32) -> u32 {
     if unsigned_add_overflows(a, b) {
         0xFFFF_FFFFu32
     } else {
-        (a + b)
+        (a + b) as u32
     }
 }
 
@@ -271,7 +271,7 @@ proof fn t514_sat_add_bounded(a: u32, b: u32)
 
 /// T514b: SatSub never goes below 0.
 pub open spec fn sat_sub_result(a: u32, b: u32) -> u32 {
-    if b > a { 0u32 } else { (a - b) }
+    if b > a { 0u32 } else { (a - b) as u32 }
 }
 
 proof fn t514_sat_sub_non_negative(a: u32, b: u32)
